Fix incorrect `padding` property value pair in `search.scss` (#1123)
This PR corrects the change to `/_sass/search.scss` made in 30290a5. This change tried to set the `padding-**top**` property to **two** values rather than set the `padding` property to these values (to represent the vertical and horizontal padding values).

I just reviewed 30290a5 and believe that this should be the other half of the fix proposed by just-the-docs/just-the-docs#1104.