add options for custom `lunr` Liquid and JS code (#1068)
This is a prototype for review and discussion. My use and testing of this PR is on top of 6da28614930108ec39065a4e15ee798aa1415c67. The changes are trival to rebase to `main` and I'm happy to do so if this prototype moves forward.


* Feature request details in linked issue, fixes just-the-docs/just-the-docs#1067
* I welcome feedback and all discussion
* A draft doc site of mine using this PR is at https://docs.hidale.com/

To use the prototype, the two include files need to be customized. Here are mine from the draft website https://github.com/diablodale/dp.docs/commit/9c0d836408af2e72dbce115a01ce6627137e66dd

Co-authored-by: Matt Wang <matt@matthewwang.me>